move surface report tests
